requirements.txt
crewai==0.130.0
crewai-tools==0.47.1
fastapi==0.110.3
uvicorn==0.29.0
python-dotenv==1.0.1
openai==1.30.5
pydantic==1.10.13
langchain==0.1.16
langchain-community==0.0.34